refactor(password-reset): abstract password reset components
